transactions-handler: filter transactions
Today we can't filter transactions to get a more precise list of them,
we can only list all transactions but without any kind of filters.

To solve that we create a method called "find" that accepts a list of
parameters exactly as the API describes to filter the transactions with
more flexibility.