fix: payment date on bulk anticipation tests
At the present moment, every acceptance test for bulk anticipations were
considering hard-coded values for the payment date. These hard-coded
instructions were in conflict with a specific business rule, that don't
allow to pick weekends as the payment date for bulk anticipations. The
problem here is that you cannot generate any builds for this project
when running the tests on mondays (lol).

The proposal for this change is to allow the tests to avoid such
conflicts by picking only work days (from monday to friday).