fix: tests that list items should be avoided
Due to recent behaviour changes on Pagar.me webservice, most of the
list-items endpoints (GET /customers, GET /transactions, etc) will
behave only as asynchronous resources. The impact on this change is
that, when querying lists of results of given data, some of the first
requests will not provide the entire list of items. With this behaviour,
we noticed in our automated tests that some assertions were crashing -
exactly the ones that creates a series of items and imeediately after
that queries for the same very items, resulting in empty results (only
for the first seconds; after a few seconds, the results were provided).

As this is a new pattern on the Pagar.me webservice, I highly encourage
that tests that query for a list of results should be avoided, from now
on.