response handler: change objects from ArrayObject to stdclass
Before this commit, the returned objects was ArrayObjects,
the problem of it is: all it sub-levels was array, and not ArrayObject.
To change this behaviour, we need a workaround, which would not be
performative.

Then, we choose to change from ArrayObjects to Array/StdClass for a better
semantics, and to divide lists from objects.